+++ lib/base/routing/router.dart
@@ -7,10 +7,10 @@ import 'package:campus_flutter/calendarComponent/views/calendars_view.dart';
 import 'package:campus_flutter/calendarComponent/views/event_creation_view.dart';
 import 'package:campus_flutter/campusComponent/screen/campus_screen.dart';
 import 'package:campus_flutter/campusComponent/screen/student_clubs_screen.dart';
-import 'package:campus_flutter/departuresComponent/views/departures_details_view.dart';
+import 'package:campus_flutter/homeComponent/view/departure/departures_details_view.dart';
 import 'package:campus_flutter/feedbackComponent/views/feedback_form_view.dart';
 import 'package:campus_flutter/feedbackComponent/views/feedback_success_view.dart';
-import 'package:campus_flutter/homeComponent/home_screen.dart';
+import 'package:campus_flutter/homeComponent/screen/home_screen.dart';
 import 'package:campus_flutter/navigaTumComponent/model/navigatum_roomfinder_map.dart';
 import 'package:campus_flutter/navigaTumComponent/views/navigatum_room_view.dart';
 import 'package:campus_flutter/onboardingComponent/views/confirm_view.dart';
