+++ lib/feedbackComponent/views/feedback_textfield.dart
@@ -1,5 +1,5 @@
 import 'package:campus_flutter/base/util/card_with_padding.dart';
-import 'package:campus_flutter/homeComponent/widgetComponent/views/widget_frame_view.dart';
+import 'package:campus_flutter/homeComponent/view/widget/widget_frame_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
