+++ lib/homeComponent/model/station.dart
