+++ lib/homeComponent/service/departures_service.dart
@@ -1,6 +1,6 @@
 import 'package:campus_flutter/base/networking/apis/mvvDeparturesApi/mvv_departures_api.dart';
 import 'package:campus_flutter/base/networking/base/rest_client.dart';
-import 'package:campus_flutter/departuresComponent/model/mvv_response.dart';
+import 'package:campus_flutter/homeComponent/model/mvv_response.dart';
 import 'package:campus_flutter/main.dart';
 
 class DeparturesService {
