+++ lib/homeComponent/view/contactCard/contact_card_error_view.dart
