+++ lib/homeComponent/view/contactCard/contact_card_loading_view.dart
