+++ lib/homeComponent/view/contactCard/link_view.dart
