+++ lib/homeComponent/view/departure/departures_details_view.dart
@@ -6,10 +6,10 @@ import 'package:campus_flutter/base/util/map_launcher.dart';
 import 'package:campus_flutter/base/util/icon_text.dart';
 import 'package:campus_flutter/base/util/last_updated_text.dart';
 import 'package:campus_flutter/base/errorHandling/error_handling_router.dart';
-import 'package:campus_flutter/departuresComponent/model/departure.dart';
-import 'package:campus_flutter/departuresComponent/model/station.dart';
-import 'package:campus_flutter/departuresComponent/viewModel/departures_viewmodel.dart';
-import 'package:campus_flutter/departuresComponent/views/departures_details_row_view.dart';
+import 'package:campus_flutter/homeComponent/model/departure.dart';
+import 'package:campus_flutter/homeComponent/model/station.dart';
+import 'package:campus_flutter/homeComponent/viewmodel/departures_viewmodel.dart';
+import 'package:campus_flutter/homeComponent/view/departure/departures_details_row_view.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
