+++ lib/homeComponent/view/departure/departures_widget_view.dart
@@ -4,12 +4,12 @@ import 'package:campus_flutter/base/util/card_with_padding.dart';
 import 'package:campus_flutter/base/util/delayed_loading_indicator.dart';
 import 'package:campus_flutter/base/errorHandling/error_handling_router.dart';
 import 'package:campus_flutter/base/routing/routes.dart';
-import 'package:campus_flutter/departuresComponent/model/departure.dart';
-import 'package:campus_flutter/departuresComponent/model/station.dart';
-import 'package:campus_flutter/departuresComponent/viewModel/departures_viewmodel.dart';
-import 'package:campus_flutter/departuresComponent/views/departures_details_row_view.dart';
-import 'package:campus_flutter/homeComponent/widgetComponent/views/preference_selection_view.dart';
-import 'package:campus_flutter/homeComponent/widgetComponent/views/widget_frame_view.dart';
+import 'package:campus_flutter/homeComponent/model/departure.dart';
+import 'package:campus_flutter/homeComponent/model/station.dart';
+import 'package:campus_flutter/homeComponent/viewmodel/departures_viewmodel.dart';
+import 'package:campus_flutter/homeComponent/view/departure/departures_details_row_view.dart';
+import 'package:campus_flutter/homeComponent/view/widget/preference_selection_view.dart';
+import 'package:campus_flutter/homeComponent/view/widget/widget_frame_view.dart';
 import 'package:campus_flutter/base/extensions/context.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
