+++ lib/homeComponent/view/widget/widget_screen.dart
@@ -2,8 +2,8 @@ import 'package:campus_flutter/base/enums/error_handling_view_type.dart';
 import 'package:campus_flutter/base/enums/home_widget.dart';
 import 'package:campus_flutter/base/util/delayed_loading_indicator.dart';
 import 'package:campus_flutter/base/errorHandling/error_handling_router.dart';
-import 'package:campus_flutter/homeComponent/widgetComponent/viewModels/home_viewmodel.dart';
-import 'package:campus_flutter/homeComponent/widgetComponent/views/home_settings_view.dart';
+import 'package:campus_flutter/homeComponent/viewmodel/home_viewmodel.dart';
+import 'package:campus_flutter/homeComponent/view/widget/home_settings_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
