+++ lib/homeComponent/viewmodel/departures_viewmodel.dart
@@ -5,11 +5,11 @@ import 'package:campus_flutter/base/enums/campus.dart';
 import 'package:campus_flutter/base/enums/user_preference.dart';
 import 'package:campus_flutter/base/util/icon_text.dart';
 import 'package:campus_flutter/base/services/location_service.dart';
-import 'package:campus_flutter/departuresComponent/model/departure.dart';
-import 'package:campus_flutter/departuresComponent/model/departures_preference.dart';
-import 'package:campus_flutter/departuresComponent/model/mvv_response.dart';
-import 'package:campus_flutter/departuresComponent/model/station.dart';
-import 'package:campus_flutter/departuresComponent/services/departures_service.dart';
+import 'package:campus_flutter/homeComponent/model/departure.dart';
+import 'package:campus_flutter/homeComponent/model/departures_preference.dart';
+import 'package:campus_flutter/homeComponent/model/mvv_response.dart';
+import 'package:campus_flutter/homeComponent/model/station.dart';
+import 'package:campus_flutter/homeComponent/service/departures_service.dart';
 import 'package:campus_flutter/main.dart';
 import 'package:campus_flutter/settingsComponent/service/user_preferences_service.dart';
 import 'package:easy_localization/easy_localization.dart';
