+++ lib/navigation_service.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:campus_flutter/base/enums/credentials.dart';
 import 'package:campus_flutter/base/routing/routes.dart';
-import 'package:campus_flutter/homeComponent/widgetComponent/views/widget_screen.dart';
+import 'package:campus_flutter/homeComponent/view/widget/widget_screen.dart';
 import 'package:campus_flutter/main.dart';
 import 'package:campus_flutter/onboardingComponent/viewModels/onboarding_viewmodel.dart';
 import 'package:campus_flutter/searchComponent/viewModels/global_search_viewmodel.dart';
