+++ lib/settingsComponent/views/appearance_settings_view.dart
@@ -5,7 +5,7 @@ import 'package:campus_flutter/base/enums/user_preference.dart';
 import 'package:campus_flutter/base/views/seperated_list.dart';
 import 'package:campus_flutter/calendarComponent/views/calendars_view.dart';
 import 'package:campus_flutter/studiesComponent/viewModel/grade_viewmodel.dart';
-import 'package:campus_flutter/homeComponent/widgetComponent/views/widget_frame_view.dart';
+import 'package:campus_flutter/homeComponent/view/widget/widget_frame_view.dart';
 import 'package:campus_flutter/main.dart';
 import 'package:campus_flutter/settingsComponent/viewModels/user_preferences_viewmodel.dart';
 import 'package:campus_flutter/settingsComponent/views/settings_view.dart';
