+++ lib/studiesComponent/view/lectureDetail/lecture_info_card_view.dart
@@ -1,6 +1,6 @@
 import 'package:campus_flutter/base/util/icon_text.dart';
 import 'package:campus_flutter/base/views/seperated_list.dart';
-import 'package:campus_flutter/homeComponent/widgetComponent/views/widget_frame_view.dart';
+import 'package:campus_flutter/homeComponent/view/widget/widget_frame_view.dart';
 import 'package:flutter/material.dart';
 
 class LectureInfoCardView extends StatelessWidget {
